remove request_helpers_inl.h
